Move event package to model
