+++ src/main/java/net/sf/jabref/gui/entryeditor/EntryEditor.java
@@ -65,7 +65,7 @@ import javax.swing.text.JTextComponent;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
-import net.sf.jabref.event.FieldChangedEvent;
+import net.sf.jabref.model.event.FieldChangedEvent;
 import net.sf.jabref.external.WriteXMPEntryEditorAction;
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.gui.EntryContainer;
