+++ src/main/java/net/sf/jabref/model/entry/BibEntry.java
@@ -36,7 +36,7 @@ import java.util.TreeSet;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
-import net.sf.jabref.event.FieldChangedEvent;
+import net.sf.jabref.model.event.FieldChangedEvent;
 import net.sf.jabref.model.FieldChange;
 import net.sf.jabref.model.database.BibDatabase;
