+++ src/test/java/net/sf/jabref/model/database/BibDatabaseTest.java
@@ -8,7 +8,7 @@ import java.util.Collections;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
-import net.sf.jabref.event.TestEventListener;
+import net.sf.jabref.model.event.TestEventListener;
 import net.sf.jabref.importer.ParserResult;
 import net.sf.jabref.importer.fileformat.BibtexParser;
 import net.sf.jabref.model.entry.BibEntry;
